Rename simulation package to app
