+++ src/META-INF/MANIFEST.MF
@@ -1,2 +1,2 @@
 Manifest-Version: 1.0
-Main-Class: simulation.Main
+Main-Class: app.Main
