+++ src/app/models/AdHocCar.java
@@ -1,4 +1,4 @@
-package simulation.models;
+package app.models;
 
 import java.util.Random;
 import java.awt.*;
