+++ src/app/models/CarQueue.java
@@ -1,4 +1,4 @@
-package simulation.models;
+package app.models;
 import java.util.LinkedList;
 import java.util.Queue;
