+++ src/app/services/Simulator.java
@@ -1,8 +1,8 @@
-package simulation.services;
+package app.services;
 
 import daytime.Daytime;
 import daytime.DaytimeSimulator;
-import simulation.views.SimulatorView;
+import app.views.SimulatorView;
 
 /**
  * Simulator contains most of the simulation logic
